Fix locking behavior.
Previous behavior allowed a SuperFigure to set itself as the current
figure before obtaining the lock.  Now, only set current figure after
getting lock.  Since this would disable the f = figure() syntax, add
a kw to constructor to set the current figure before grabbing the
lock.  Use this with caution.

Additionally, the stand-along functions need to check if plotting is
locked before checking current fig.  If it is locked by the same
thread, (or not locked, but there is a current figure, meaning the
f = figure() syntax is being used), then use the current figure.
Otherwise, create a new figure after getting lock.  This behavior
required the switch from Lock to RLock.  This also fixes the hang
on nested with figure()s, but that syntax still may not work.