Mess with reinteract_output to improve with blocks.
Add a functions so that reinteract_output can be passed in to the module
from the worksheet.  Then, SuperFigure will call reinteract_output(self)
from __exit__, so you don't have to return the figure yourself.
Additionally, it molests reinteract_output during the with block to
suppress output.

This is only as a test.  It works only in a single worksheet at a time,
and the things it does to reinteract_output should not be done in polite
company.