Add printing support.
Added a print_result() method to SuperFigure.  Largely uses code
from refigure.  When using GTKCairo, redraw the figure onto the
printing context.  Otherwise, save the figure as a bitmap and use
that.  (usetex is incompatible with cairo, for example.)  The
calling function already uses context.save() and restore(), so we
don't have to do it here.