Mess with reinteract_output through the worksheet frame.
Use inspect.currentframe() and frame.f_back to find the worksheet
frame.  Then, get reinteract_output from there, turn it off during
the with block, and use it to output the figure at the end.  The
keyword argument disable_output is added to SuperFigure to govern
whether to disable reinteract_output (defaults to True).

This is still likely fragile and probably not a good idea.