Don't create FigureCanvasGTK* until create_widget()
Previously, the canvas had been created in SuperFigure.__init__.  However,
this runs in one of the worker threads, not the mainloop thread.  You're
not supposed to do GTK things outside of the mainloop thread.  In practice,
the only thing that screwed up was that signal handlers were lost when the
widget was unrealized.  An important signal handler drew the figure on
expose-events.  Therefore, the figure would not be redrawn after an
unrealize/realize cycle, for example, after a text reflow.  This is what
caused bug #1.

However, if the user wants to call savefig(), there'd better be a canvas,
so that the figure can be drawn.  So we start off with a FigureCanvasBase
in __init__, as it provides the print_figure and various print_* methods.
Then, in create_widget, we switch to the appropriate GTK canvas.

This exposes another bug - a gtk.Toolbar (used in the NavigationToolbar2)
produces ref_count warnings and errors when it is unrealized.