Add rcParams settings for refigure2
Added rcParams['refigure.printdpi' and 'refigure.disableoutput'] to
set the DPI when printing with the bitmap backend and the default
for disabling the output from with blocks.  Also will read an rc
file named refigurerc in the same directory as the module file and
set parameters from there.  Should add more places to look for the
rc file.