Allow the figure to be controlled by the keyboard
This functionality is handled by a FigureManager in matplotlib,
so we add the Base one.  Events are automatically connected, but
we need to give the figure focus on being clicked and ensure that
keyboard events don't bubble up to the TextView.

Also added a rc setting, refigure.keyboardcontrol, for whether
this should be enabled.  Some users might not like the worksheet
loosing focus whenever a figure is clicked on.