Resize sidebar figures to fit sidebar width
We adjust the figure size through the DPI, leaving the nominal size in
inches unchanged.  This way, the layout of the figure doesn't change as
the sidebar is resized.  We're assuming that users have chosen refigure2
because they're making complicated figures, so we don't want to mess
them up when resizing the sidebar.