Take the printing DPI to be 72, ignoring context.get_dpi_x()
Windows sometimes gives a DPI of 600, or similar, which leads to really
mis-sized plots.  See Reinteract commit e3644466.

We also make sure cdpi is a float, since dpi may not be and the two are
involved in a division in the bitmap printing mechanism.

Thanks to Richard Thomas for trouble-shooting.